Alter get_variables to return an abstract Mapping.
This allows a simpler, more expressive way of getting the variables that
are available on a component, as well as asking for a specific variable
if you already know which one you want. Implementations that are asking
for information on variables over-the-wire can return an object that
executes the required queries on-demand.